+++ console/src/project/list/index.js
@@ -13,18 +13,26 @@ import {getCurrentOrg} from "../../session/reducers";
 const {actions, reducers, selectors, sagas} = DataItem("project/list", [], api.listProjects);
 
 const columns = [
+    {key: "visibility", label: "Access", collapsing: true},
     {key: "name", label: "Project", collapsing: true},
     {key: "description", label: "Description"}
 ];
 
 const projectNameKey = "name";
+const visibilityKey = "visibility";
 
 const cellFn = (row, key) => {
+    if (key === visibilityKey) {
+        const n = row[key];
+        return <Icon name={n === "PUBLIC" ? "unlock" : "lock"} color={n === "PRIVATE" ? "red" : "grey"}/>;
+    }
+
     // link to the project page
     if (key === projectNameKey) {
         const n = row[key];
         return <Link to={`/project/${n}`}>{n}</Link>;
     }
+
     return row[key];
 };